Moved views from components folder to a views folder
